replace devshell with mkShell
devshell doesn't work with nixpkgs's C infrastructure as it doesn't
provide the right environment variables like
NIX_CFLAGS_COMPILE/NIX_LDFLAGS. This is a long-standing issue.
Also export LD_LIBRARY_PATH is not a good idea as it will crash other
applications that are not part of the devshell i.e. editors.