remove configuration file
This is something users run once per machine, so they won't likely have
their dotfiles in the installer they are running this from and cli flags
seem more convinient.
Not having configuration files has the advantage that given a bug report
we can tell how they run nixos-facter without also having to ask for
configuration. If we ever get to the point where commandline flags
becomes no longer feasible we can still add a configuration format back.
But at the moment it doesn't look like this. This also saves us a bunch
of dependency that no longer need to be compiled by users of the tool.