feat: refactor HardwareDevice model
Removes custom `MarshalJSON` to make docs easier.

As a consequence, we no longer hoist the `Detail` field in the JSON output for `base_class.value == 257`. Instead, I've made improvements to ensure fields that don't make sense are omitted.

This was only really happening for the Memory hardware class, as we do something custom for CPU which was the original reason we did this.

I've also removed some fields which were internal or not providing useful info.

BREAKING CHANGE: removes some fields and changes behaviour related to omitting.

Signed-off-by: Brian McGee <brian@bmcgee.ie>