fix: classification of pci and usb devices
Improves overall classification by capturing the class list of the underlying hardware item.

When the overall hardware class is unknown, we call back to the class list which typically identifies it as usb or pci etc.

Adds the `class` and `class_list` fields to the JSON output for clarity.

We should consider whether we allow the same hardware item in different sections of the hardware report based on the `class_list`.

Signed-off-by: Brian McGee <brian@bmcgee.ie>
Co-authored-by: Jörg Thalheim <Mic92@users.noreply.github.com>
Signed-off-by: Brian McGee <brian@bmcgee.ie>