feat: replace custom virt logic with systemd-detect-virt
We already have a dependency on `systemdMinimal` through `hwinfo`, so rather than duplicating the logic, we should just lean into it for now.

If we eventually move to a static binary, then we can revisit this.

NOTE: when next releasing to nixpkgs, we will need to update its runtime inputs.

Closes #142

Signed-off-by: Brian McGee <brian@bmcgee.ie>