feat: capture touchpads
Touchpads are not captured by `hwinfo`.

We read `/proc/bus/input/devices` and `/run/udev/data` to find them and add corresponding `HardwareDevice` entries.

Note: `attached_to` is not currently supported, and will be set to 0. It's not obvious the logic for resolving the report entries that a touchpad is connected to.
Signed-off-by: Brian McGee <brian@bmcgee.ie>